requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
aiosqlite==0.19.0
pytz==2024.1
aiohttp==3.9.3
pandas==2.2.0
pytest==8.0.0
pytest-asyncio==0.23.0
redis==5.0.1
apscheduler==3.10.4
tenacity==8.2.3